refactor: Move agent creation code to a separate agents.py file
